+++ src/cipher/secp256k1-go/secp256k1-go2/xy.go
@@ -163,7 +163,7 @@ func (xy *XY) SetXYZ(a *XYZ) {
 	xy.Y = a.Y
 }
 
-func (xy *XY) precomp(w int) (pre []XY) {
+func (xy *XY) precomp(w int) (pre []XY) { // nolint: unused,megacheck
 	pre = make([]XY, (1 << (uint(w) - 2)))
 	pre[0] = *xy
 	var X, d, tmp XYZ