+++ src/daemon/gateway_test.go
@@ -223,7 +223,6 @@ func TestGateway_GetWallets(t *testing.T) {
 		name            string
 		enableWalletAPI bool
 		wallets         wallet.Wallets
-		getWalletError  error
 		err             error
 	}{
 		{